refactor: migrate group/clan to lazy_container
- Add lazy_container with ensure, ensure_guard, and negative cache
- Move group/clan container into entity headers and split actions/events
- Replace service::group/clan with server groups/clans containers
- Fix shard_container async try_enter miss handling and factory tombstones
- Fix thread retry loop capture and use read lock before HTTP in actions
- Route AMQP updated_clan through apply_updated
