+++ server/game/src/handler/amqp/broadcast_clan.cpp
@@ -12,6 +12,6 @@ async::task<void> broadcast_clan::handle(const internal_resp::BroadcastClan& mes
     if (message.host == fb::config<uint32_t>("id"))
         co_return;
 
-    co_await this->server.clan.on_error(message.error);
-    co_await this->server.clan.on_broadcast(message.clan, message.message, message.type);
+    co_await this->server.clans.on_error(message.error);
+    co_await this->server.clans.on_broadcast(message.clan, message.message, message.type);
 }
