+++ server/game/src/handler/amqp/destroy_group.cpp
@@ -12,6 +12,6 @@ async::task<void> destroy_group::handle(const internal_resp::DestroyGroup& messa
     if (message.host == fb::config<uint32_t>("id"))
         co_return;
 
-    co_await this->server.group.on_error(message.error, message.actor.name);
-    co_await this->server.group.on_destroyed(message.actor.name, message.group_id);
+    co_await this->server.groups.on_error(message.error, message.actor.name);
+    co_await this->server.groups.on_destroyed(message.actor.name, message.group_id);
 }
